add express code

+++ controllers/clothingItems.js
@@ -1,5 +1,5 @@
 const ClothingItem = require("../models/clothingItem");
-const { NOT_FOUND, BAD_REQUEST, DEFAULT } = require("../utils/errors");
+const { NOT_FOUND, BAD_REQUEST, DEFAULT, FORBIDDEN_ERROR } = require("../utils/errors");
 
 const createItem = (req, res) => {
   console.log(req);
@@ -14,7 +14,7 @@ const createItem = (req, res) => {
     })
     .catch((error) => {
       if (error.name === "ValidationError") {
-        return res.status(BAD_REQUEST).send({message: "Invalid data" });
+        return res.status(BAD_REQUEST).send({ message: "Invalid data" });
       }
       return res.status(DEFAULT).send({ message: "Error from  createItem" });
     });
@@ -43,18 +43,28 @@ const getItems = (req, res) => {
 
 const deleteItem = (req, res) => {
   const { itemId } = req.params;
-
-  console.log(itemId);
-  ClothingItem.findByIdAndDelete(itemId)
+  //we need to check if the currently logged in user owns this item and is allowed to delete it
+  ClothingItem.findById(itemId)
     .orFail()
-    .then((items) =>
-      res.status(200).send({ items })
-    )
+    .then((item) => {
+      if (req.user._id !== item.owner.toString()) {
+        const error = new Error();
+        error.name = "ForbiddenError";
+        throw error;
+      }
+      return ClothingItem.findByIdAndDelete(itemId);
+    })
+    .then((items) => res.status(200).send({ items }))
     .catch((err) => {
       if (err.name === "DocumentNotFoundError") {
         return res.status(NOT_FOUND).send({ message: "item not found" });
-      } if (err.name === "CastError") {
+      }
+      if (err.name === "CastError") {
         return res.status(BAD_REQUEST).send({ message: "item request failed" });
+      }
+      //if block for the forbidden error
+      if (err.name === "ForbiddenError"){
+        return res.status(FORBIDDEN_ERROR).send({ message: "forbidden action" });
       }
       return res.status(DEFAULT).send({ message: "item request error" });
     });
@@ -70,9 +80,12 @@ const likeItem = (req, res) => {
     .then((items) => res.status(200).send(items))
     .catch((err) => {
       if (err.name === "DocumentNotFoundError") {
-        return res.status(NOT_FOUND).send({ message: "items not found"  });
-      } if (err.name === "CastError") {
-        return res.status(BAD_REQUEST).send({ message: "items request failed" });
+        return res.status(NOT_FOUND).send({ message: "items not found" });
+      }
+      if (err.name === "CastError") {
+        return res
+          .status(BAD_REQUEST)
+          .send({ message: "items request failed" });
       }
       return res.status(DEFAULT).send({ message: "Get items failed" });
     });
@@ -89,8 +102,11 @@ const deleteLike = (req, res) => {
     .catch((err) => {
       if (err.name === "DocumentNotFoundError") {
         return res.status(NOT_FOUND).send({ message: "items not found" });
-      } if (err.name === "CastError") {
-        return res.status(BAD_REQUEST).send({ message: "items request failed" });
+      }
+      if (err.name === "CastError") {
+        return res
+          .status(BAD_REQUEST)
+          .send({ message: "items request failed" });
       }
       return res.status(DEFAULT).send({ message: "Get items failed" });
     });
